+++ client/src/components/SearchBarPlanets/index.jsx
@@ -1,0 +1,48 @@
+import React, {useState} from "react";
+
+const SearchBar =() => {
+    const [searchPlanet, setSearchPlanet] = useState('');
+
+    const handleInputChange = (event) => {
+        setSearchPlanet(event.target.value);
+    };
+
+    const handleSearch = (event) => {
+        event.preventDefault();
+        //search logic
+        console.log('Searching for', searchPlanet);
+    };
+
+    return (
+        <form onSubmit={ handleSearch}>
+            <input
+                type="text"
+                value = {searchPlanet}
+                onChange={handleInputChange}
+                placeholder="Search"
+                />
+                <button type="submit">Search</button>
+        </form>
+    );
+};
+
+export default SearchBar;
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
