+++ client/src/main.jsx
@@ -10,7 +10,7 @@ import './index.css';
 
 // importing all pages that the router will use for front end
 
-// import App from './App';
+import App from './App';
 // import AreaInfo from './pages/AreaInfo/AreaInfo';
 // import Home from './pages/Home/Home';
 // import Login from './pages/Login/Login';
@@ -20,11 +20,11 @@ import './index.css';
 
 // Accessible routes need to be defined, as well as which components will respond to which URLs
 
-// const router = createBrowserRouter([
-//   {
-//     path: '/',
-//     element: <App />,
-//     children: [
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <App />,
+    children: [
 //       {
 //         index: true,
 //         element: <Home />,
@@ -49,9 +49,9 @@ import './index.css';
 //         path: '/areainfo',
 //         element: <AreaInfo />,
 //       },
-//     ],
-//   },
-// ]);
+    ],
+  },
+]);
 
 // Render the Router Provider component
 ReactDOM.createRoot(document.getElementById('root')).render(
