+++ client/src/pages/Profile/Profile.jsx
@@ -1,5 +1,7 @@
 import { useQuery } from '@apollo/client';
 
+import SearchBarArea from '../../components/SearchBarArea';
+import SearchBarPlanet from '../../components/SearchBarPlanets';
 
 
 
@@ -15,6 +17,5 @@ import { useQuery } from '@apollo/client';
 
 
 
-<div>
-    
-</div>
+
+
